@@ -1,32 +1,32 @@
 name: domain
 description: Domain module
 publish_to: 'none' 
-version: 1.0.0
-author:
+version: 0.0.1
 homepage: ' '
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.0.2"
+  sdk: ">=2.16.0 <3.0.0"
+  flutter: "2.10.4"
 
 dependencies:
+  device_info_plus: 3.2.2
   flutter:
     sdk: flutter
-  
   foundation:
     path: ../foundation
-  
-  freezed_annotation: 0.14.1
-  get_it: 6.0.0
-  shared_preferences: 2.0.5
+  fpdart: 0.0.14
+  freezed_annotation: 0.15.0
+  get_it: 7.2.0
+  project_linter:
+    path: ../project_linter
+  shared_preferences: 2.0.13
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-    
-  build_runner: 1.12.2
-  freezed: 0.14.1+1
-  mockito: 5.0.3
-
+  analyzer: 2.5.0
+  build_runner: 2.1.7
+  freezed: 0.15.0+1
+  mocktail: 0.2.0
+  platform: 3.1.0
+  test: ^1.20.2
 
 flutter:
