@@ -1,21 +1,25 @@
 name: foundation
-description: Here you can add common utilities
-version: 1.0.1
-author:
+description: foundation
+version: 0.0.1
 homepage: ' '
 
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.2.2"
+  sdk: ">=2.16.0 <3.0.0"
+  flutter: "2.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  fpdart: 0.0.14
+  logging: 1.0.2
+  project_linter:
+    path: ../project_linter
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  platform: 3.1.0
 
 flutter:
