@@ -1,49 +1,47 @@
 name: presentation
 description: Presentation module
-version: 1.0.0
-author:
-homepage: ' '
-
+version: 0.0.1
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.2.2"
+  sdk: ">=2.16.0 <3.0.0"
+  flutter: "2.10.4"
 
 dependencies:
-  flutter:
-    sdk: flutter
-
-  flutter_localizations:
-    sdk: flutter
-
-  foundation:
-    path: ../foundation
+  adaptive_theme: 2.3.1
+  bloc_concurrency: 0.2.0
+  cupertino_icons: 1.0.4
   domain:
     path: ../domain
+  flutter:
+    sdk: flutter
+  flutter_bloc: 8.0.1
+  flutter_localizations:
+    sdk: flutter
+  flutter_svg: 1.0.3
+  foundation:
+    path: ../foundation
+  fpdart: 0.0.14
+  freezed_annotation: 0.15.0
+  get_it: 7.2.0
+  google_fonts: 2.3.1
+  intl: 0.17.0
+  project_linter:
+    path: ../project_linter
   remote:
     path: ../remote
 
-  cupertino_icons: ^1.0.2
-  flutter_dotenv: 4.0.0-nullsafety.0
-  flutter_bloc: 7.0.0
-  freezed_annotation: 0.14.1
-  get_it: 6.0.0
-  google_fonts: 2.0.0
-  intl: 0.17.0
-  stacked_themes: 0.3.0
-
 dev_dependencies:
+  analyzer: 2.5.0
+  bloc_test: 9.0.3
+  build_runner: 2.1.7
   flutter_test:
     sdk: flutter
-
-  analyzer: 1.3.0
-  bloc_test: 8.0.0
-  build_runner: 1.12.2
-  freezed: 0.14.1+1
-  mockito: 5.0.3
-  pedantic: 1.11.0
-
+  freezed: 0.15.0+1
+  lints: 1.0.1
+  mockingjay: 0.3.0
+  mocktail: 0.3.0
+  platform: 3.1.0
 
 flutter:
   uses-material-design: true
