@@ -1,34 +1,35 @@
 name: remote
 description: Remote module
 publish_to: 'none' 
-version: 1.0.0
-author:
+version: 0.0.1
 homepage: ' '
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.2.2"
+  sdk: ">=2.16.0 <3.0.0"
+  flutter: "2.10.0"
 
 dependencies:
-  flutter:
-    sdk: flutter
-  
-  foundation:
-    path: ../foundation
+  collection: 1.15.0
+  dio: 4.0.6
   domain:
     path: ../domain
-
-  retrofit: 2.0.0-beta1
-  json_annotation: 4.0.1
-  json_serializable: 4.1.0
-  pretty_dio_logger: 1.2.0-beta-1
+  flutter:
+    sdk: flutter
+  foundation:
+    path: ../foundation
+  fpdart: 0.0.14
+  get_it: 7.2.0
+  json_annotation: 4.4.0
+  json_serializable: 6.1.5
+  project_linter:
+    path: ../project_linter
+  retrofit: 3.0.1+1
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  
-  build_runner: 1.12.2
-  mockito: 5.0.3
-  retrofit_generator: ^2.0.0-beta3
+  build_runner: 2.1.8
+  mocktail: 0.3.0
+  platform: 3.1.0
+  retrofit_generator: 4.0.1
+  test: 1.20.2
 
 flutter:
