@@ -1,8 +1,11 @@
-class Insets {
-  static const double xSmall = 4;
-  static const double small = 8;
+abstract class Insets {
+  const Insets._();
+
+  static const double small = 4;
+  static const double xSmall = 8;
   static const double medium = 12;
   static const double large = 16;
   static const double xLarge = 24;
   static const double xxLarge = 32;
+  static const double xxxLarge = 42;
 }
