@@ -2,5 +2,5 @@ part of 'dashboard_bloc.dart';
 
 @freezed
 class DashboardEvent with _$DashboardEvent {
-  const factory DashboardEvent.started() = _Started;
+  const factory DashboardEvent.started() = _DashboardStarted;
 }
