@@ -1,15 +1,47 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:presentation/application/dimen.dart';
+import 'package:presentation/common/state_type.dart';
+import 'package:presentation/screens/dashboard/bloc/dashboard_bloc.dart';
+import 'package:presentation/widgets/empty_page_widget.dart';
+import 'package:presentation/widgets/error_page_widget.dart';
+import 'package:presentation/widgets/loading_page_widget.dart';
 
 class DashboardScreen extends StatelessWidget {
   static const String routeName = '/dashboard';
 
+  // context.strings.dashboardScreenTitle,
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: Container(
-        color: Colors.white,
-        child: const Center(
-          child: FlutterLogo(),
+      appBar: AppBar(
+        title: const Text('Dashboard'),
+      ),
+      body: SafeArea(
+        top: false,
+        child: BlocBuilder<DashboardBloc, DashboardState>(
+          builder: (context, state) {
+            return state.type.map(
+              loading: () => const LoadingPage(),
+              loaded: () {
+                return Padding(
+                  padding: const EdgeInsets.symmetric(
+                    vertical: Insets.medium,
+                    horizontal: Insets.large,
+                  ),
+                  child: Column(
+                    children: [
+                      Text(state.user?.name ?? ''),
+                      Text(state.user?.email ?? ''),
+                    ],
+                  ),
+                );
+              },
+              empty: () => const EmptyPage(),
+              error: () => const ErrorPage(),
+            );
+          },
         ),
       ),
     );
